+++ crates/volta-migrate/src/lib.rs
@@ -16,12 +16,14 @@ mod v1;
 mod v2;
 mod v3;
 mod v4;
+mod v5;
 
 use v0::V0;
 use v1::V1;
 use v2::V2;
 use v3::V3;
 use v4::V4;
+use v5::V5;
 
 use log::{debug, info};
 use volta_core::error::Fallible;
@@ -42,6 +44,7 @@ enum MigrationState {
     V2(Box<V2>),
     V3(Box<V3>),
     V4(Box<V4>),
+    V5(Box<V5>),
 }
 
 /// Macro to simplify the boilerplate associated with detecting a tagged state.
@@ -81,7 +84,13 @@ macro_rules! detect_tagged {
     }
 }
 
-detect_tagged!((v4, V4, V4), (v3, V3, V3), (v2, V2, V2), (v1, V1, V1));
+detect_tagged!(
+    (v5, V5, V5),
+    (v4, V4, V4),
+    (v3, V3, V3),
+    (v2, V2, V2),
+    (v1, V1, V1)
+);
 
 impl MigrationState {
     fn current() -> Fallible<Self> {
@@ -162,12 +171,13 @@ fn detect_and_migrate() -> Fallible<()> {
     // latest version. We then apply the migrations sequentially here: V0 -> V1 -> ... -> VX
     loop {
         state = match state {
-            MigrationState::Empty(e) => MigrationState::V3(Box::new(e.try_into()?)),
+            MigrationState::Empty(e) => MigrationState::V5(Box::new(e.try_into()?)),
             MigrationState::V0(zero) => MigrationState::V1(Box::new((*zero).try_into()?)),
             MigrationState::V1(one) => MigrationState::V2(Box::new((*one).try_into()?)),
             MigrationState::V2(two) => MigrationState::V3(Box::new((*two).try_into()?)),
             MigrationState::V3(three) => MigrationState::V4(Box::new((*three).try_into()?)),
-            MigrationState::V4(current) => {
+            MigrationState::V4(four) => MigrationState::V5(Box::new((*four).try_into()?)),
+            MigrationState::V5(current) => {
                 debug!(
                     "Volta directory is already using the current layout at {}",
                     current.home.root().display()